# metal : reduce command encoding overhead